Added Reschedule button funcionallity.
