+++ app/src/main/java/org/eyeseetea/malariacare/layout/adapters/survey/PlannedAdapter.java
@@ -19,17 +19,12 @@
 
 package org.eyeseetea.malariacare.layout.adapters.survey;
 
-import android.app.DatePickerDialog;
-import android.app.Dialog;
 import android.content.Context;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
-import android.widget.Button;
-import android.widget.DatePicker;
-import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
@@ -40,20 +35,15 @@ import org.eyeseetea.malariacare.R;
 import org.eyeseetea.malariacare.database.model.Program;
 import org.eyeseetea.malariacare.database.model.Survey;
 import org.eyeseetea.malariacare.database.utils.PreferencesState;
-import org.eyeseetea.malariacare.database.utils.Session;
 import org.eyeseetea.malariacare.database.utils.planning.PlannedHeader;
 import org.eyeseetea.malariacare.database.utils.planning.PlannedItem;
-import org.eyeseetea.malariacare.database.utils.planning.PlannedItemBuilder;
 import org.eyeseetea.malariacare.database.utils.planning.PlannedSurvey;
+import org.eyeseetea.malariacare.database.utils.planning.ScheduleListener;
 import org.eyeseetea.malariacare.database.utils.planning.SurveyPlanner;
-import org.eyeseetea.malariacare.utils.Constants;
 import org.eyeseetea.malariacare.utils.AUtils;
+import org.eyeseetea.malariacare.utils.Constants;
 
-import java.text.DateFormat;
-import java.util.Calendar;
-import java.util.Date;
 import java.util.List;
-import java.util.Locale;
 
 /**
  * Created by arrizabalaga on 14/09/15.
@@ -277,7 +267,7 @@ public class PlannedAdapter extends BaseAdapter {
         //ScheduledDate
         textView=(TextView)rowLayout.findViewById(R.id.planning_survey_schedule_date);
         textView.setText(AUtils.formatDate(plannedSurvey.getNextAssesment()));
-        textView.setOnClickListener(new ScheduleListener(plannedSurvey.getSurvey()));
+        textView.setOnClickListener(new ScheduleListener(plannedSurvey.getSurvey(),context));
 
         //background color
         int colorId=plannedSurvey.getPlannedHeader().getSecondaryColor();
@@ -354,82 +344,6 @@ public class PlannedAdapter extends BaseAdapter {
         public void onClick(View v) {
             toggleSection(plannedHeader);
         }
-    }
-
-    class ScheduleListener implements View.OnClickListener {
-        Survey survey;
-        Date newScheduledDate;
-
-        ScheduleListener(Survey survey){this.survey=survey;}
-
-        @Override
-        public void onClick(View v){
-            final Dialog dialog = new Dialog(context);
-            dialog.setContentView(R.layout.planning_schedule_dialog);
-            dialog.setTitle(R.string.planning_title_dialog);
-
-            //Set current date
-            final Button scheduleDatePickerButton=(Button)dialog.findViewById(R.id.planning_dialog_picker_button);
-            scheduleDatePickerButton.setText(AUtils.formatDate(survey.getScheduledDate()));
-            //On Click open an specific DatePickerDialog
-            scheduleDatePickerButton.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    //Init secondary datepicker with current date
-                    Calendar calendar = Calendar.getInstance();
-                    if(survey.getScheduledDate()!=null){
-                        calendar.setTime(survey.getScheduledDate());
-                    }
-                    //Show datepickerdialog -> updates newScheduledDate and button
-                    new DatePickerDialog(PlannedAdapter.this.context, new DatePickerDialog.OnDateSetListener() {
-
-                        public void onDateSet(DatePicker view, int year, int monthOfYear, int dayOfMonth) {
-                            Calendar newCalendar = Calendar.getInstance();
-                            newCalendar.set(year, monthOfYear, dayOfMonth);
-                            newScheduledDate = newCalendar.getTime();
-                            scheduleDatePickerButton.setText(AUtils.formatDate(newScheduledDate));
-                        }
-
-                    },calendar.get(Calendar.YEAR), calendar.get(Calendar.MONTH), calendar.get(Calendar.DAY_OF_MONTH)).show();
-                }
-            });
-
-            //Listens to close button
-            Button dialogButton = (Button) dialog.findViewById(R.id.planning_dialog_close_button);
-            dialogButton.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    dialog.dismiss();
-                }
-            });
-
-            dialogButton = (Button) dialog.findViewById(R.id.planning_dialog_ok_button);
-            dialogButton.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    //TODO do something
-                    //Check fields are ok
-                    String comment = ((EditText) dialog.findViewById(R.id.planning_dialog_comment)).getText().toString();
-                    if (!validateFields(newScheduledDate, comment)) {
-                        return;
-                    }
-                    //Reschedule survey
-                    survey.reschedule(newScheduledDate, comment);
-                    //Recalculate items
-                    reloadItems(PlannedItemBuilder.getInstance().buildPlannedItems());
-
-                    dialog.dismiss();
-                }
-            });
-
-            dialog.show();
-        }
-
-        private boolean validateFields(Date newDate,String comment){
-            return newDate!=null;
-        }
-
-
     }
 
 }
